flask/requirements.txt
flask==3.0.2
flask-socketio==5.3.6
flask-cors==4.0.0
python-socketio==5.11.1
python-engineio==4.9.0
eventlet==0.35.2
redis==5.0.1
mysql-connector-python==8.3.0
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
flask-login==0.5.0
werkzeug>=3.0.0
dnspython==2.1.0
pysnmp==4.4.12
paramiko==2.8.0
reportlab==3.6.8
PyJWT==2.8.0 